Move detect_distribution and dependents to mkosi.backend
