+++ mkosi/__init__.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import argparse
-import ast
 import base64
 import configparser
 import contextlib
@@ -14,7 +13,6 @@ import dataclasses
 import datetime
 import errno
 import fcntl
-import functools
 import getpass
 import glob
 import hashlib
@@ -79,6 +77,7 @@ from mkosi.backend import (
     Verb,
     add_packages,
     chown_to_running_user,
+    detect_distribution,
     die,
     disable_pam_securetty,
     is_centos_variant,
@@ -164,14 +163,6 @@ DRACUT_SYSTEMD_EXTRAS = [
 
 
 T = TypeVar("T")
-V = TypeVar("V")
-
-
-def dictify(f: Callable[..., Iterator[Tuple[T, V]]]) -> Callable[..., Dict[T, V]]:
-    def wrapper(*args: Any, **kwargs: Any) -> Dict[T, V]:
-        return dict(f(*args, **kwargs))
-
-    return functools.update_wrapper(wrapper, f)
 
 
 def list_to_string(seq: Iterator[str]) -> str:
@@ -180,29 +171,6 @@ def list_to_string(seq: Iterator[str]) -> str:
     ['a', "b", 11] → "'a', 'b', 11"
     """
     return str(list(seq))[1:-1]
-
-@dictify
-def read_os_release() -> Iterator[Tuple[str, str]]:
-    try:
-        filename = "/etc/os-release"
-        f = open(filename)
-    except FileNotFoundError:
-        filename = "/usr/lib/os-release"
-        f = open(filename)
-
-    with f:
-        for line_number, line in enumerate(f, start=1):
-            line = line.rstrip()
-            if not line or line.startswith("#"):
-                continue
-            m = re.match(r"([A-Z][A-Z_0-9]+)=(.*)", line)
-            if m:
-                name, val = m.groups()
-                if val and val[0] in "\"'":
-                    val = ast.literal_eval(val)
-                yield name, val
-            else:
-                print(f"{filename}:{line_number}: bad line {line!r}", file=sys.stderr)
 
 
 def print_running_cmd(cmdline: Iterable[str]) -> None:
@@ -5208,38 +5176,6 @@ def parse_bytes(num_bytes: Optional[str], *, sector_size: int = 512) -> int:
         result += sector_size - rem
 
     return result
-
-
-def detect_distribution() -> Tuple[Optional[Distribution], Optional[str]]:
-    try:
-        os_release = read_os_release()
-    except FileNotFoundError:
-        return None, None
-
-    dist_id = os_release.get("ID", "linux")
-    dist_id_like = os_release.get("ID_LIKE", "").split()
-    version = os_release.get("VERSION", None)
-    version_id = os_release.get("VERSION_ID", None)
-    version_codename = os_release.get("VERSION_CODENAME", None)
-    extracted_codename = None
-
-    if version:
-        # extract Debian release codename
-        m = re.search(r"\((.*?)\)", version)
-        if m:
-            extracted_codename = m.group(1)
-
-    d: Optional[Distribution] = None
-    for the_id in [dist_id, *dist_id_like]:
-        d = Distribution.__members__.get(the_id, None)
-        if d is not None:
-            break
-
-    if d in {Distribution.debian, Distribution.ubuntu} and (version_codename or extracted_codename):
-        # debootstrap needs release codenames, not version numbers
-        version_id = version_codename or extracted_codename
-
-    return d, version_id
 
 
 def remove_glob(*patterns: PathString) -> None:
